+++ src/main/java/XXLChess/Bishop.java
@@ -2,7 +2,7 @@ package XXLChess;
 
 public class Bishop extends Piece {
 
-  public Bishop(int x, int y, String code, Square curSquare) {
+  public Bishop(float x, float y, String code, Square curSquare) {
     super(x, y, code, curSquare);
     this.value = 3.625 * (this.isWhite ? 1 : -1);
   }
