+++ src/main/java/XXLChess/Camel.java
@@ -1,7 +1,7 @@
 package XXLChess;
 
 public class Camel extends Piece {
-  public Camel(int x, int y, String code, Square curSquare) {
+  public Camel(float x, float y, String code, Square curSquare) {
     super(x, y, code, curSquare);
     this.value = 2 * (this.isWhite ? 1 : -1);
   }
