+++ src/main/java/XXLChess/Chancellor.java
@@ -1,7 +1,7 @@
 package XXLChess;
 
 public class Chancellor extends Piece {
-  public Chancellor(int x, int y, String code, Square curSquare) {
+  public Chancellor(float x, float y, String code, Square curSquare) {
     super(x, y, code, curSquare);
     this.value = 8.5 * (this.isWhite ? 1 : -1);
   }
