+++ src/main/java/XXLChess/Guard.java
@@ -2,7 +2,7 @@ package XXLChess;
 
 public class Guard extends Piece {
 
-  public Guard(int x, int y, String code, Square curSquare) {
+  public Guard(float x, float y, String code, Square curSquare) {
     super(x, y, code, curSquare);
     this.value = 5 * (this.isWhite ? 1 : -1);
   }
