+++ src/main/java/XXLChess/Queen.java
@@ -1,7 +1,7 @@
 package XXLChess;
 
 public class Queen extends Piece {
-  public Queen(int x, int y, String code, Square curSquare) {
+  public Queen(float x, float y, String code, Square curSquare) {
     super(x, y, code, curSquare);
     this.value = 9.5 * (this.isWhite ? 1 : -1);
   }
